components/backend/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
importlib-metadata==6.8.0
platformdirs==3.10.0
psycopg==3.1.10
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
yapf==0.40.1
zipp==3.16.2
gunicorn==21.2.0
